grub: retitle grml boot menu entries
The default menu entry now just reads "<grml-name> <version>", so for
example "grml-small-arm64 2024.12". The additional options are in a
menu entry just titled " -> Options".

To allow for removing the name/version for the options menu, grub.cfg
now sorts the %SHORTNAME%_options.cfg always directly after the
%SHORTNAME%_default.cfg.

Use unicode.pf2 to allow for Unicode characters in the Options and
Addons menu entries.

# Conflicts:
#	grml-live