scripts/forensic-mark-readonly: Fix properly finding the parent of a device-mapper device
Parent devices of device-mapper devices use a directory named "holders" in
their sysfs block device directory that lists all device-mapper devices
which use that device. This is unlike partition devices of block devices
which have directory directly in the sysfs block device directory. So if
the current device name is not found directly in the sysfs block device
directory, then check the "holders" directory.

Signed-off-by: Glenn Washburn <development@efficientek.com>