scripts/forensic-mark-readonly: Fix not setting to readonly when no parent
When parent_device is empty, `grep -qw "${parent_device}"` will always
return true regardless of input. So when READONLY_IGNORE is set and no
parent device is found for a device, that device is never marked readonly.
However, newly seen devices should be marked as readonly unless otherwise
specified (eg. in READONLY_IGNORE). Ensure that devices with no parent are
marked as readonly.

Signed-off-by: Glenn Washburn <development@efficientek.com>