isolinux: Drop bootoption "gmt"
We replaced "gmt" with "utc" some time ago, see: 65ee43c.

In the meanwhile we changed the timezone and localtime handling and
dropped support for the bootoption "utc", as it is the default nowadays,
see: grml/grml-autoconfig@e96da70 + grml/grml-live@34167e2f.

TZ defaults to UTC and the RTC defaults to be set to UTC.

Adjusted wording for "localtime" in grml-cheatcodes.txt too.

Related: grml/grml-live#296, grml/grml-live#243