Announce SSH service with Avahi
We want to announce the SSH service with Avahi.

The avahi-daemon is shipped in GRML_FULL but needs to be enabled.
Grml boots into the grml-boot.target but avahi-daemon starts in
multi-user.target by default, we need to override that.

Also, an Avahi ssh service configuration file needs to be deployed.
The avahi-ssh.service deploys the SSH service config for Avahi when the
SSH service is started. The service avahi-ssh.service needs to be
enabled too.

To discover the SSH service in the local network the following command
can be used:

  avahi-browse -d local _ssh._tcp --resolve -t

Issue: grml/grml-live#73