Create GRML_GHACI class as normal file
Move the file from an inline creation in test/gha-build-iso.sh. Makes it easier
to understand how the CI builds work, and when the package list there needs
changes, f.e. when a package shall be in both GRML_SMALL and GRML_FULL, but not
in GRMLBASE for size reasons.