grml-live: deprecate setting individual *_OUTPUT vars
I intend to make changes later to how and where these variables are setup and
passing all of them through to the python tools is annoying.

I can see how it was useful to move individual output directories in the past,
but I'm counting on computers being fast and big enough today.