grml-live: deprecate FAI_DEBOOTSTRAP
Instead support a new BOOTSTRAP_MIRROR variable, which has *just* the mirror
URL.

Judging by an example in the docs, since introducing minifai, we were not fully
compatible with the old syntax.

Ideally minifai would determine the mirror URL from the class config, but that
needs some form of marking one of the sources files as the bootstrap one. Ideas
welcome.