new SSHGEN class to generate SSH keys at build time
This class enables the SSH service and generates SSH keys for the same
reason as the TOR class generating onion service keys: those are shown
in the build logs and can be used to authenticate the remote host (a
redundant measure to the onion service name, of course).

It also enables the SSH service, roughly equivalent to passing the
`ssh` boot option.

Finally, another shim is required here to inject a valid SSH public
key in the image, so you can login over SSH. In my case, I have an
extra class that only has this one script which does:

    gpg --export-ssh-key anarcat@debian.org | tee -a $target/root/.ssh/authorized_keys

This could also be folded in the SSHGEN class, but I'm not sure how to
do variables yet, so that's not yet standardized.