explain operators how to complete the TOR configuration
Here we take the approach of not doing anything fancy with SSH out of
the ordinary: we tell grml to boot with the ssh option and pass a
secret password or a netconfig= fetch at build time.

We pick this approach instead of adding another class because we can't
quite tell how to easily pass a SSH public key to an eventual SSHGEN
class.