doc: simplify "local-files" example with -I option
I have found that section to be difficult to read, at best. I had to
read the code block a few times to even understand what it did - why
do we need GRML_FAI_CONFIG in here?

Assuming all this is unnecessary, it seems much easier to do overlays
directly on the grml-live commandline with -I.

So let's show that!

We still refer to the CHROOT_INSTALL (configuration) variable there so
power users can write that in their configuration files, of course.