Random password generation: ignore write errors from tr
tr(1) reads from /dev/urandom and outputs an infinite stream of
alphanumeric characters, but head -c 28 reads exactly 28 bytes and
exists. When head exits, the pipe closes but tr tries to write more data
to the now-closed pipe and receives a SIGPIPE signal. This causes tr to
report the "Broken pipe" error before terminating:

  [  OK  ] Bootoption ssh found, trying to set password for root and user grml
  tr: write error: Broken pipe

Thanks: crass for the bug report
Closes: https://github.com/grml/grml/issues/250