Rework password handling to unify random password generation
The code in config_vnc() was another implementation of the random
password generation, while we need it for all the boot options passwd,
ssh + vnc.

Unify the random password generation into a single function which
returns the password on stdout, and use it within the passwd, ssh + vnc
functions. Rename the function name from set_passwd to
get_random_password to better reflect what's going on.